move i18nProvider back to index.js in separate component with App.js
